ATR-828: moved Code Map specific graph semantic info into subfolder
