+++ src/Acuminator/Acuminator.Vsix/Tool Windows/CodeMap/Model/GraphSemanticModelForCodeMap.cs
@@ -10,6 +10,7 @@ using Microsoft.CodeAnalysis;
 using Acuminator.Utilities.Common;
 using Acuminator.Utilities.Roslyn.Semantic;
 using Acuminator.Utilities.Roslyn.Semantic.PXGraph;
+using Acuminator.Vsix.ToolWindows.CodeMap.Graph;
 
 namespace Acuminator.Vsix.ToolWindows.CodeMap
 {
